Expose language-specific naming utilities in Bazel for downstream code generators
