+++ src/google/protobuf/compiler/cpp/BUILD.bazel
@@ -8,6 +8,15 @@ load("@rules_proto//proto:defs.bzl", "proto_library")
 load("//build_defs:cpp_opts.bzl", "COPTS")
 
 cc_library(
+    name = "names",
+    hdrs = ["names.h"],
+    copts = COPTS,
+    include_prefix = "google/protobuf/compiler/cpp",
+    visibility = ["//visibility:public"],
+    deps = ["//src/google/protobuf:protobuf_nowkt"],
+)
+
+cc_library(
     name = "cpp",
     srcs = [
         "enum.cc",
@@ -38,7 +47,6 @@ cc_library(
         "message.h",
         "message_field.h",
         "message_layout_helper.h",
-        "names.h",
         "options.h",
         "padding_optimizer.h",
         "parse_function_generator.h",
@@ -53,6 +61,7 @@ cc_library(
         "//src/google/protobuf/compiler:__pkg__",
     ],
     deps = [
+        ":names",
         "//src/google/protobuf:protobuf_nowkt",
         "//src/google/protobuf/compiler:code_generator",
         "@com_google_absl//absl/base:core_headers",
