+++ src/google/protobuf/compiler/csharp/BUILD.bazel
@@ -7,6 +7,15 @@ load("@rules_pkg//:mappings.bzl", "pkg_files", "strip_prefix")
 load("//build_defs:cpp_opts.bzl", "COPTS")
 
 cc_library(
+    name = "csharp_names",
+    hdrs = ["csharp_names.h"],
+    copts = COPTS,
+    include_prefix = "google/protobuf/compiler/csharp",
+    visibility = ["//visibility:public"],
+    deps = ["//src/google/protobuf:protobuf_nowkt"],
+)
+
+cc_library(
     name = "csharp",
     srcs = [
         "csharp_doc_comment.cc",
@@ -36,7 +45,6 @@ cc_library(
         "csharp_map_field.h",
         "csharp_message.h",
         "csharp_message_field.h",
-        "csharp_names.h",
         "csharp_options.h",
         "csharp_primitive_field.h",
         "csharp_reflection_class.h",
@@ -56,6 +64,7 @@ cc_library(
         "//src/google/protobuf/compiler:__pkg__",
     ],
     deps = [
+        ":csharp_names",
         "//src/google/protobuf:protobuf_nowkt",
         "//src/google/protobuf/compiler:code_generator",
         "@com_google_absl//absl/container:flat_hash_set",
