+++ src/google/protobuf/compiler/php/php_names.h
@@ -31,8 +31,6 @@
 #ifndef GOOGLE_PROTOBUF_COMPILER_PHP_GENERATOR_H__
 #define GOOGLE_PROTOBUF_COMPILER_PHP_GENERATOR_H__
 
-#include "google/protobuf/compiler/code_generator.h"
-#include "google/protobuf/compiler/php/php_names.h"
 #include "google/protobuf/descriptor.h"
 
 #include <string>
@@ -44,37 +42,12 @@ namespace protobuf {
 namespace compiler {
 namespace php {
 
-struct Options;
-
-class PROTOC_EXPORT Generator : public CodeGenerator {
- public:
-  virtual bool Generate(
-      const FileDescriptor* file,
-      const std::string& parameter,
-      GeneratorContext* generator_context,
-      std::string* error) const override;
-
-  bool GenerateAll(const std::vector<const FileDescriptor*>& files,
-                   const std::string& parameter,
-                   GeneratorContext* generator_context,
-                   std::string* error) const override;
-
-  uint64_t GetSupportedFeatures() const override {
-    return FEATURE_PROTO3_OPTIONAL;
-  }
-
- private:
-  bool Generate(
-      const FileDescriptor* file,
-      const Options& options,
-      GeneratorContext* generator_context,
-      std::string* error) const;
-};
-
-inline bool IsWrapperType(const FieldDescriptor* descriptor) {
-  return descriptor->cpp_type() == FieldDescriptor::CPPTYPE_MESSAGE &&
-      descriptor->message_type()->file()->name() == "google/protobuf/wrappers.proto";
-}
+// To skip reserved keywords in php, some generated classname are prefixed.
+// Other code generators may need following API to figure out the actual
+// classname.
+PROTOC_EXPORT std::string GeneratedClassName(const Descriptor* desc);
+PROTOC_EXPORT std::string GeneratedClassName(const EnumDescriptor* desc);
+PROTOC_EXPORT std::string GeneratedClassName(const ServiceDescriptor* desc);
 
 }  // namespace php
 }  // namespace compiler
